refactor(validation): split the git layer and its tests under the size ceiling
Pure move, no behavior change. Round 8 pushed
`check_index_line_endings.py` to 513 lines and
`test_check_index_line_endings_repo.py` to 505, both past the 500-line
`file-size` ceiling, so the count ratchet would have blocked the push.

`index_line_endings_git.py` takes the subprocess boundary: how git is invoked
(`git_environment`, `run_git`, `run_git_paths`, `has_commits`) and whether the
running git can answer at all (`git_version`, `require_attr_source`). Nothing
in it knows what a line-ending violation is. The gate keeps the policy: which
scopes to read, what to report, when to refuse to write, and how to remediate.

`test_check_index_line_endings_env.py` takes the matching test seam. The repo
module asks whether a tracked blob contradicts its attributes; the env module
asks whether the gate is looking at the repository it was given, judging by
the attributes that repository stores, on a git that can answer.

Four tests reached the moved symbols through the gate's namespace and now
name the module that owns them.

Refs #5475
